refactor: move save/load to new fileio module
