+++ smolmodels/__init__.py
@@ -1,3 +1,3 @@
 from .models import Model as Model
-from .models import load_model as load_model
-from .models import save_model as save_model
+from .fileio import load_model as load_model
+from .fileio import save_model as save_model
